refactor: define layouts by roles
