+++ resources/js/Layouts/BaseLayout.vue
@@ -6,13 +6,33 @@ import Banner from '@/components/Common/Banner.vue'
 import {ref, watch} from 'vue'
 import {type PageProps} from '@/Types/PageProps'
 
-const props = defineProps<{ pages: Page[] } & PageProps>()
+const props = defineProps<PageProps>()
 const status = ref<string | undefined>(props.flash.status)
 
 watch(() => props.flash, flash => {
   status.value = flash.status
 }, { immediate: true })
 
+const pages: Page[] = []
+
+if (!!props.user?.isAdmin || props.user?.isSuperAdmin) {
+  pages.push(
+    { title: 'Testy', href: '/admin/quizzes' },
+    { title: 'Szkoły', href: '/admin/schools' },
+    { title: 'Uczniowie', href: '/admin/users' },
+  )
+}
+else {
+  pages.push({ title: 'Konkursy', href: '/dashboard' })
+}
+
+pages.push(
+  { title: 'Profil', href: '/profile' },
+)
+
+if (props.user?.isSuperAdmin) {
+  pages.push({ title: 'Administratorzy', href: '/admin/admins'})
+}
 </script>
 
 <template>
